requirements.txt
pandas>=2.1
httpx>=0.27
python-dotenv>=1.0
typer>=0.12
pydantic>=2.7
shapely>=2.0
pytest>=8.0
pytest-asyncio>=0.23
pytest-httpx>=0.30
googlemaps>=4.10.0